// starter-frontend/src/app/api/generated/model/language-enum.ts
/**
 * Starter API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `en` - English * `fr` - French * `nl` - Dutch * `it` - Italy * `es` - Spain
 */
export enum LanguageEnumDto {

    En = 'en',

    Fr = 'fr',

    Nl = 'nl',

    It = 'it',

    Es = 'es'
}